bugfix: Conditionally build bond extension based on ruby version
Before, it was erroneously performing the condition check in the source
gemfile.  This condition is evaluated while building the package, and
that result is used when installing the gem: therefore, if you built the
gem with ruby-1.9.2, but installed with ruby-1.8.7, it wouldn't try and
build the extension.

With this fix, the decision to build the extension (or not) is made
while trying to build the gem.  Unfortunately, I don't believe there's a
better way as this seems to be the first code rubygems runs on the
target machine while installing a gem.

Fixes http://github.com/cldwalker/bond/issues/issue/5